EcsProject/requirements.txt
aws-cdk-lib==2.17.0
constructs>=10.0.0,<11.0.0

aws_cdk.aws_certificatemanager